完成pmem_read和pmem_write函数，此外发现inst 和 pc的更新存在时序问题，pc是上升沿更新，inst 是下降沿更新，现在先尝试分别更新，下一步就是纠正时序逻辑

+++ npc/csrc/monitor/sdb/expr.cpp
@@ -327,7 +327,7 @@ uint32_t eval(int p, int q, bool* success) {
 				if (tokens[op].type == TK_DEREF) {
 					*success = true;
 					// return vaddr_read(val, 4);
-                    return pmem_read(val);
+                    return (uint32_t)pmem_read(val);
 				}
 				if (tokens[op].type == TK_SUBTRACT){
 					*success = true;
